backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pydantic[email]
pydantic-settings
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
slowapi
